requirements.txt
absl-py==1.4.0
accelerate==0.25.0
attrs==22.2.0
cachetools==5.3.0
# certifi @ file:///croot/certifi_1671487769961/work/certifi
charset-normalizer==3.1.0
cloudpickle==2.2.1
colorama==0.4.6
cycler==0.11.0
deepdiff==6.3.0
dm-env==1.6
dm-env-rpc==1.1.5
dm-tree==0.1.8
docopt==0.6.2
enum34==1.1.10
exceptiongroup==1.1.1
fonttools==4.38.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.16.3
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.59.0
grpcio==1.51.3
gym==0.26.2
gym-notices==0.0.8
idna==3.4
imageio==2.26.1
immutabledict==2.2.3
importlib-metadata==6.1.0
iniconfig==2.0.0
joblib==1.2.0
jsonpickle==1.5.2
kiwisolver==1.4.4
kmeans-pytorch
lbforaging==1.1.1
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.5.3
mock==5.0.1
mpyq==0.2.5
munch==2.5.0
numpy
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
ordered-set==4.1.0
packaging==23.0
pandas==1.3.5
Pillow==9.4.0
pluggy==1.0.0
portpicker==1.5.2
probscale==0.2.5
protobuf==3.19.5
psutil==5.9.4
py-cpuinfo==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygame==2.3.0
pyglet==2.0.7
pyparsing==3.0.9
PySC2==4.0.0
pytest==7.2.2
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML
requests==2.32.3
requests-oauthlib==1.3.1
rsa==4.9
s2clientprotocol==5.0.11.89720.0
s2protocol==5.0.11.89720.0
sacred==0.8.2
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.2
six==1.16.0
sk-video==1.1.10
SMAC @ git+https://github.com/oxwhirl/smac.git@8a092e4bd0c6f5d3cf929523c67e1bca861463aa
smmap==5.0.0
snakeviz==2.1.1
tensorboard
tensorboard-data-server==0.6.1
tensorboard-logger==0.1.0
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tomli==2.0.1
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tornado==6.2
transformers == 4.39.1
tqdm
typing_extensions==4.5.0
urllib3==1.26.15
websocket-client==1.5.1
Werkzeug==2.2.3
wrapt==1.15.0
zipp==3.15.0
tensorboardX
SMACv2 @ git+https://github.com/oxwhirl/smacv2.git
wandb[media]